+++ callsignlookuptools/__info__.py
@@ -3,7 +3,7 @@
 
 
 __project__ = "callsignlookuptools"
-__summary__ = "QRZ and callook.info API interfaces in Python with sync and async support"
+__summary__ = "QRZ, Callook, HamQTH, and QRZCQ API interface with sync and async support"
 __webpage__ = "https://callsignlookuptools.miaow.io"
 
 __version__ = "0.0.0"
@@ -12,3 +12,4 @@ __author__ = "classabbyamp, 0x5c"
 __email__ = "dev@kb6.ee, dev@0x5c.io"
 
 __license__ = "BSD 3-Clause"
+__copyright__ = "2021"