+++ callsignlookuptools/qrzcq/qrzcqsync.py
@@ -1,31 +1,31 @@
 """
-qrztools: synchronous editon
+qrzcqtools: synchronous editon
 ---
 Copyright 2021 classabbyamp, 0x5c
 Released under the terms of the BSD 3-Clause license.
 """
 
 
-from typing import Optional
 from urllib.parse import urlencode
+from typing import Optional
 
 import requests
 
 from ..common import mixins, dataclasses, exceptions
 from ..common.constants import DEFAULT_USERAGENT
-from .qrz import QrzClientAbc
+from .qrzcq import QrzCqClientAbc
 
 
-class QrzSyncClient(mixins.SyncXmlAuthMixin, mixins.SyncMixin, QrzClientAbc):
-    """Synchronous QRZ API client
+class QrzCqSyncClient(mixins.SyncXmlAuthMixin, mixins.SyncMixin, QrzCqClientAbc):
+    """Synchronous QRZCQ API client
 
-    :param username: QRZ username
+    :param username: QRZCQ username
     :type username: str
-    :param password: QRZ password
+    :param password: QRZCQ password
     :type password: str
-    :param session_key: QRZ login session key
+    :param session_key: QRZCQ login session key
     :type session_key: str
-    :param useragent: Useragent for QRZ
+    :param useragent: Useragent for QRZCQ
     :type useragent: str
     :param session: A requests session to use for requests
     :type session: requests.Session
@@ -44,7 +44,8 @@ class QrzSyncClient(mixins.SyncXmlAuthMixin, mixins.SyncMixin, QrzClientAbc):
             raise exceptions.CallsignLookupError("Invalid Callsign")
         try:
             self._check_session(
-                s=self._session_key
+                s=self._session_key,
+                agent=self._useragent
             )
         except exceptions.CallsignLookupError:
             self._login(
@@ -57,12 +58,13 @@ class QrzSyncClient(mixins.SyncXmlAuthMixin, mixins.SyncMixin, QrzClientAbc):
             query=callsign.upper(),
             resp=self._do_query(
                 s=self._session_key,
-                callsign=callsign.upper()
+                callsign=callsign.upper(),
+                agent=self._useragent
             )
         )
 
     def _do_query(self, **query) -> bytes:
         with self._session.get(self._base_url + urlencode(query)) as resp:
             if resp.status_code != 200:
-                raise exceptions.CallsignLookupError(f"Unable to connect to QRZ (HTTP Error {resp.status_code})")
+                raise exceptions.CallsignLookupError(f"Unable to connect to QRZCQ (HTTP Error {resp.status_code})")
             return resp.content