+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for QRZ lookup.
 - Support for Callook lookup.
+- Support for HamQTH lookup.
+- Support for QRZCQ lookup.
 - Documentation.
 - Command-line interface.
 
 
 <!--
-[Unreleased]: https://github.com/miaowware/callsignlookuptools/compare/1.0.0...HEAD
-[1.0.0]: https://github.com/miaowware/callsignlookuptools/releases/tag/1.0.0
+[Unreleased]: https://github.com/miaowware/callsignlookuptools/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/miaowware/callsignlookuptools/releases/tag/v1.0.0
 -->